+++ src/main/java/com/nanonaitor/arsenal/network/ModNetwork.java
@@ -10,7 +10,8 @@ import net.minecraftforge.network.PacketDistributor;
 import net.minecraftforge.network.SimpleChannel;
 
 public final class ModNetwork {
-    public static final byte FLAIL = 1, BALL_CHAIN = 2, RAM = 3, BULWARK_BASH = 4, CLAW = 5;
+    public static final byte FLAIL = 1, BALL_CHAIN = 2, RAM = 3, BULWARK_BASH = 4,
+        CLAW = 5, CLAW_MAIN = 6;
     private static SimpleChannel channel;
 
     public static void init() {